refactor demo-rollup/bench
