+++ examples/demo-rollup/src/lib.rs
@@ -1,4 +1,2 @@
 #![deny(missing_docs)]
 #![doc = include_str!("../README.md")]
-/// A very simple example of rollup configuration.
-pub mod rng_xfers;
